refactor: rename DispatcherRepository to DispatcherProvider
