+++ app/src/main/java/com/example/nimblesurveys/di/UtilModules.kt
@@ -1,9 +1,9 @@
 package com.example.nimblesurveys.di
 
 import com.example.nimblesurveys.data.repository.TimeRepositoryImpl
-import com.example.nimblesurveys.domain.repository.DispatcherRepository
+import com.example.nimblesurveys.domain.provider.DispatcherProvider
 import com.example.nimblesurveys.domain.repository.TimeRepository
-import com.example.nimblesurveys.util.DispatcherRepositoryImpl
+import com.example.nimblesurveys.util.DispatcherProviderImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -22,5 +22,5 @@ object TimeModule {
 object DispatcherModule {
 
     @Provides
-    fun provideDispatcherRepository(): DispatcherRepository = DispatcherRepositoryImpl()
+    fun provideDispatcherProvider(): DispatcherProvider = DispatcherProviderImpl()
 }
