+++ app/src/main/java/com/example/nimblesurveys/util/DispatcherProviderImpl.kt
@@ -1,9 +1,9 @@
 package com.example.nimblesurveys.util
 
-import com.example.nimblesurveys.domain.repository.DispatcherRepository
+import com.example.nimblesurveys.domain.provider.DispatcherProvider
 import kotlinx.coroutines.Dispatchers
 
-class DispatcherRepositoryImpl: DispatcherRepository {
+class DispatcherProviderImpl: DispatcherProvider {
     override fun io() = Dispatchers.IO
     override fun main() = Dispatchers.Main
 }
